move states into common directory
