+++ fala/apps/adviser/forms.py
@@ -3,7 +3,7 @@
 from django import forms
 from django.utils.translation import gettext_lazy as _
 from django.utils.safestring import mark_safe
-from .regions import Region
+from fala.apps.adviser.regions import Region
 from fala.apps.adviser.utils import validate_postcode_and_return_country
 import laalaa.api as laalaa
